impel changes

+++ apps/health_campaign_field_worker_app/lib/blocs/household_overview/household_overview.dart
@@ -4,7 +4,6 @@ import 'dart:async';
 import 'package:collection/collection.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:path/path.dart';
 
 import '../../models/data_model.dart';
 import '../../utils/typedefs.dart';
@@ -22,6 +21,7 @@ class HouseholdOverviewBloc
   final ProjectBeneficiaryDataRepository projectBeneficiaryRepository;
   final TaskDataRepository taskDataRepository;
   final SideEffectDataRepository sideEffectDataRepository;
+  final ReferralDataRepository referralDataRepository;
 
   HouseholdOverviewBloc(
     super.initialState, {
@@ -31,6 +31,7 @@ class HouseholdOverviewBloc
     required this.householdMemberRepository,
     required this.taskDataRepository,
     required this.sideEffectDataRepository,
+    required this.referralDataRepository,
   }) {
     on(_handleDeleteHousehold);
     on(_handleDeleteIndividual);
@@ -160,6 +161,13 @@ class HouseholdOverviewBloc
           tasks.map((e) => e.clientReferenceId).whereNotNull().toList(),
     ));
 
+    final referrals = await referralDataRepository.search(ReferralSearchModel(
+      projectBeneficiaryClientReferenceId: projectBeneficiaries
+          .map((e) => e.clientReferenceId)
+          .whereNotNull()
+          .toList(),
+    ));
+
     // Update the state with the loaded data and stop loading.
     emit(
       state.copyWith(
@@ -170,6 +178,7 @@ class HouseholdOverviewBloc
           tasks: tasks.isEmpty ? null : tasks,
           projectBeneficiaries: projectBeneficiaries,
           sideEffects: sideEffects,
+          referrals: referrals,
         ),
         loading: false,
       ),
@@ -182,16 +191,68 @@ class HouseholdOverviewBloc
     HouseholdOverviewEmitter emit,
   ) async {
     // Delete the household from the repository.
+    final HouseholdModel? existingHousehold =
+        (await householdRepository.search(HouseholdSearchModel(
+      clientReferenceId: [event.householdModel.clientReferenceId],
+    )))
+            .firstOrNull;
+
     await householdRepository.delete(
       event.householdModel.copyWith(
-        rowVersion: event.householdModel.rowVersion,
+        id: existingHousehold?.id,
+        rowVersion: existingHousehold?.rowVersion ?? 1,
+        nonRecoverableError: existingHousehold?.nonRecoverableError ?? false,
       ),
     );
 
     // Iterate through individual members of the household.
     for (final i in event.members) {
       // Delete the individual from the repository.
-      await individualRepository.delete(i);
+      final IndividualModel? existingIndividual =
+          (await individualRepository.search(IndividualSearchModel(
+        clientReferenceId: [i.clientReferenceId],
+      )))
+              .firstOrNull;
+      await individualRepository.delete(i.copyWith(
+        id: existingIndividual?.id,
+        rowVersion: existingIndividual?.rowVersion ?? 1,
+        nonRecoverableError: existingIndividual?.nonRecoverableError ?? false,
+      ));
+
+      if (event.projectBeneficiaryType == BeneficiaryType.individual) {
+        // Search for project beneficiary associated with the deleted individual.
+        final projectBeneficiaries = await projectBeneficiaryRepository.search(
+          ProjectBeneficiarySearchModel(
+            beneficiaryClientReferenceId: [
+              i.clientReferenceId,
+            ],
+          ),
+        );
+
+        // Delete the associated project beneficiaries.
+        for (final projectBeneficiary in projectBeneficiaries) {
+          await projectBeneficiaryRepository.delete(
+            projectBeneficiary.copyWith(
+              rowVersion: projectBeneficiary.rowVersion,
+              clientAuditDetails: (projectBeneficiary
+                              .clientAuditDetails?.createdBy !=
+                          null &&
+                      projectBeneficiary.clientAuditDetails?.createdTime !=
+                          null)
+                  ? ClientAuditDetails(
+                      createdBy:
+                          projectBeneficiary.clientAuditDetails!.createdBy,
+                      createdTime:
+                          projectBeneficiary.clientAuditDetails!.createdTime,
+                      lastModifiedBy:
+                          projectBeneficiary.clientAuditDetails!.lastModifiedBy,
+                      lastModifiedTime: DateTime.now().millisecondsSinceEpoch,
+                    )
+                  : null,
+            ),
+          );
+        }
+      }
 
       // Search for household members associated with the deleted individual.
       final householdMember =
@@ -210,12 +271,23 @@ class HouseholdOverviewBloc
       }
     }
 
-    // Delete the project beneficiary associated with the household.
-    await projectBeneficiaryRepository.delete(
-      event.projectBeneficiaryModel.copyWith(
-        rowVersion: event.projectBeneficiaryModel.rowVersion,
-      ),
-    );
+    if (BeneficiaryType.household == event.projectBeneficiaryType) {
+      // Delete the project beneficiary associated with the household.
+      final ProjectBeneficiaryModel? existingProjectBeneficiary =
+          (await projectBeneficiaryRepository
+                  .search(ProjectBeneficiarySearchModel(
+        clientReferenceId: [event.projectBeneficiaryModel.clientReferenceId],
+      )))
+              .firstOrNull;
+      await projectBeneficiaryRepository.delete(
+        event.projectBeneficiaryModel.copyWith(
+          id: existingProjectBeneficiary?.id,
+          rowVersion: existingProjectBeneficiary?.rowVersion ?? 1,
+          nonRecoverableError:
+              existingProjectBeneficiary?.nonRecoverableError ?? false,
+        ),
+      );
+    }
   }
 
   // This function handles the deletion of an individual from a household.
@@ -224,7 +296,49 @@ class HouseholdOverviewBloc
     HouseholdOverviewEmitter emit,
   ) async {
     // Delete the individual from the repository.
-    await individualRepository.delete(event.individualModel);
+    final IndividualModel? existingIndividual =
+        (await individualRepository.search(IndividualSearchModel(
+      clientReferenceId: [event.individualModel.clientReferenceId],
+    )))
+            .firstOrNull;
+    await individualRepository.delete(event.individualModel.copyWith(
+      id: existingIndividual?.id,
+      rowVersion: existingIndividual?.rowVersion ?? 1,
+      nonRecoverableError: existingIndividual?.nonRecoverableError ?? false,
+    ));
+
+    if (event.projectBeneficiaryType == BeneficiaryType.individual) {
+      // Search for project beneficiary associated with the deleted individual.
+      final projectBeneficiaries = await projectBeneficiaryRepository.search(
+        ProjectBeneficiarySearchModel(
+          beneficiaryClientReferenceId: [
+            event.individualModel.clientReferenceId,
+          ],
+        ),
+      );
+
+      // Delete the associated project beneficiaries.
+      for (final projectBeneficiary in projectBeneficiaries) {
+        await projectBeneficiaryRepository.delete(
+          projectBeneficiary.copyWith(
+            rowVersion: projectBeneficiary.rowVersion,
+            clientAuditDetails: (projectBeneficiary
+                            .clientAuditDetails?.createdBy !=
+                        null &&
+                    projectBeneficiary.clientAuditDetails?.createdTime != null)
+                ? ClientAuditDetails(
+                    createdBy: projectBeneficiary.clientAuditDetails!.createdBy,
+                    createdTime:
+                        projectBeneficiary.clientAuditDetails!.createdTime,
+                    lastModifiedBy:
+                        projectBeneficiary.clientAuditDetails!.lastModifiedBy,
+                    lastModifiedTime: DateTime.now().millisecondsSinceEpoch,
+                  )
+                : null,
+          ),
+        );
+      }
+    }
 
     // Search for household members associated with the deleted individual.
     final householdMembers = await householdMemberRepository.search(
